refactor: rename feature-task-list-ui module to feature-task-ui
