+++ android/app/build.gradle.kts
@@ -54,7 +54,7 @@ android {
 dependencies {
     implementation(project(":android:android-core"))
     implementation(project(":android:feature:feature-stopwatch-ui"))
-    implementation(project(":android:feature:feature-task-list-ui"))
+    implementation(project(":android:feature:feature-task-ui"))
     implementation(project(":android:feature:feature-settings-ui"))
     implementation(project(":kmp:shared"))
     implementation(project(":kmp:core:core-shared"))
