--- android/feature/feature-task-list-ui/src/main/AndroidManifest.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest package="com.akjaw.timi.android.feature.task.list.ui">
-
-</manifest>
