+++ android/feature/feature-task-ui/src/main/java/com/akjaw/timi/android/feature/task/ui/list/TaskListScreen.kt
@@ -1,4 +1,4 @@
-package com.akjaw.timi.android.feature.task.list.ui.view
+package com.akjaw.timi.android.feature.task.ui.list
 
 import androidx.compose.animation.animateContentSize
 import androidx.compose.animation.core.Spring
@@ -48,7 +48,7 @@ import com.akjaw.timi.android.core.view.theme.TimiComposeTheme
 import com.akjaw.timi.android.core.view.theme.taskShape
 import com.akjaw.timi.android.core.view.theme.taskTextColorFor
 import com.akjaw.timi.android.core.view.toComposeColor
-import com.akjaw.timi.android.feature.task.list.ui.R
+import com.akjaw.timi.android.feature.task.ui.R
 import com.akjaw.timi.kmp.feature.task.api.list.domain.model.Task
 import com.akjaw.timi.kmp.feature.task.api.list.presentation.TaskListViewModel
 import org.koin.androidx.compose.get
