+++ settings.gradle
@@ -7,7 +7,7 @@ rootProject.name = "Timi-Compose"
 include ':android:app'
 include ':android:android-core'
 include ':android:feature:feature-settings-ui'
-include ':android:feature:feature-task-list-ui'
+include ':android:feature:feature-task-ui'
 include ':android:feature:feature-stopwatch-ui'
 include ':kmp:shared'
 include ':kmp:core:core-shared'
